requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
langchain-groq==0.1.9
langchain-core==0.2.38
xgboost==2.1.1
numpy==1.26.4
scikit-learn==1.5.1
httpx==0.27.2
